Sports classification merge to main

+++ backend/spring-boot-jwt-authentication/src/main/java/com/irojas/demojwt/Auth/ModelDTO/UserDTO.java
@@ -10,17 +10,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserDTO {
-    int id;
+    Long id;
 	String email;
     String firstname;
     String lastname;
     String country;
     
     
-    public int getId() {
+    public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
